Minor improvements

+++ elf/src/main/java/com/ledmington/elf/FileHeader.java
@@ -243,7 +243,7 @@ public final class FileHeader {
      *
      * @return The size of this file header.
      */
-    public int getHeaderSize() {
+    public short getHeaderSize() {
         return headerSize;
     }
